close F-3 (sim pkg-fresh drift check) + F-9 (Kotlin hops type fidelity) from pass-17 (#91)
F-3 (HIGH): sim/check-pkg-fresh.sh's interface diff was vacuous as wired into CI. Both
workflows run sim/build-wasm.sh (which overwrites the working-tree sim/pkg from HEAD)
immediately before the guard, so the guard's `diff fresh-build vs sim/pkg/$f` compared a
fresh build against that same just-rebuilt copy: it could never detect a stale commit
(proven in the pass-17 audit). Fixed: the interface diff now reads the COMMITTED blob via
`git show HEAD:sim/pkg/$f`, so it catches an export/interface change that was never rebuilt
into the committed pkg, regardless of step order. Proven: added a temporary wasm export,
ran build-wasm.sh (the exact CI order), and the fixed guard reports DRIFT where the old one
said "fresh"; reverted. The wire-stamp cross-check already read committed; a purely-
behavioral non-wire change lives only in the excluded nondeterministic .wasm and is
exercised by scenario-check on a fresh build (documented in the comment).

F-9 (LOW): the Kotlin SDK's HopMessage.hops was `Byte` where the C ABI is uint8_t and the
UniFFI driver bindings expose the equivalent field as UByte, so a hop count >= 128 would
render negative. Changed the public field to UByte (recovered from the FFI Byte via
.toUByte() at construction; the FFI boundary keeps Byte for correct JNA marshalling of the
native uint8_t). Updated hashCode + the value-type tests. Kotlin SDK tests + the 80%
coverage gate pass.


GitOrigin-RevId: be9202e85b3a7914f32fa49260d446bbd0f0778b

+++ sdk/wrappers/kotlin/src/main/kotlin/sh/hop/Hop.kt
@@ -24,7 +24,11 @@ enum class HopRole(val c: Int) { DIALER(0), ACCEPTOR(1) }
  *  that might mutate them (a `data class` can't return defensive copies from its generated accessors).
  *  equals/hashCode are content-based (value semantics); as with any value carrying a mutable array, do
  *  not mutate a field and then rely on it as a HashMap/HashSet key. */
-data class HopMessage(val from: ByteArray, val contentType: String, val body: ByteArray, val hops: Byte, val createdAt: Long) {
+// F-9: `hops` is `UByte`, not `Byte`. The C ABI field is uint8_t (0..255) and the UniFFI-generated
+// driver bindings expose the equivalent field as UByte; a signed Byte would render a hop count >= 128
+// as a negative number. The FFI boundary itself (InboxSink.invoke, hop_message_status) keeps Byte for
+// correct JNA marshalling of a native uint8_t; we reinterpret to UByte at this public surface.
+data class HopMessage(val from: ByteArray, val contentType: String, val body: ByteArray, val hops: UByte, val createdAt: Long) {
     /** A defensive copy of the sender address (mutate this freely without affecting the message). */
     fun fromCopy(): ByteArray = from.copyOf()
     /** A defensive copy of the body bytes (mutate this freely without affecting the message). */
@@ -40,7 +44,7 @@ data class HopMessage(val from: ByteArray, val contentType: String, val body: By
         var r = from.contentHashCode()
         r = 31 * r + contentType.hashCode()
         r = 31 * r + body.contentHashCode()
-        r = 31 * r + hops
+        r = 31 * r + hops.toInt()
         r = 31 * r + createdAt.hashCode()
         return r
     }
@@ -284,7 +288,7 @@ class HopNode private constructor(rawPtr: Pointer) : AutoCloseable {
                 from = from?.getByteArray(0, 32) ?: ByteArray(32),
                 contentType = ct ?: "",
                 body = body?.getByteArray(0, blen.toInt()) ?: ByteArray(0),
-                hops = hops, createdAt = created))
+                hops = hops.toUByte(), createdAt = created))
         }, null)
     }
